requirements.txt
# Data processing & ML
pandas
numpy
scikit-learn
joblib
psycopg2-binary

# Text preprocessing (choose one)
nltk
pymorphy3
pymorphy3-dicts-ru

# FastAPI web service
fastapi
uvicorn[standard]
pydantic
jinja2

# Progress bar
tqdm

#Deployment
gunicorn

#Frontend testing
streamlit
